+++ libraries/common/src/main/java/androidx/media3/common/FileTypes.java
@@ -57,6 +57,7 @@ public final class FileTypes {
    *   <li>{@link #PNG}
    *   <li>{@link #WEBP}
    *   <li>{@link #BMP}
+   *   <li>{@link #HEIF}
    * </ul>
    */
   @Documented
@@ -64,7 +65,7 @@ public final class FileTypes {
   @Target(TYPE_USE)
   @IntDef({
     UNKNOWN, AC3, AC4, ADTS, AMR, FLAC, FLV, MATROSKA, MP3, MP4, OGG, PS, TS, WAV, WEBVTT, JPEG,
-    MIDI, AVI, PNG, WEBP, BMP
+    MIDI, AVI, PNG, WEBP, BMP, HEIF
   })
   public @interface Type {}
 
@@ -131,6 +132,9 @@ public final class FileTypes {
   /** File type for the BMP format. */
   public static final int BMP = 19;
 
+  /** File type for the HEIF format. */
+  public static final int HEIF = 20;
+
   @VisibleForTesting /* package */ static final String HEADER_CONTENT_TYPE = "Content-Type";
 
   private static final String EXTENSION_AC3 = ".ac3";
@@ -170,6 +174,7 @@ public final class FileTypes {
   private static final String EXTENSION_WEBP = ".webp";
   private static final String EXTENSION_BMP = ".bmp";
   private static final String EXTENSION_DIB = ".dib";
+  private static final String EXTENSION_HEIC = ".heic";
 
   private FileTypes() {}
 
@@ -241,6 +246,8 @@ public final class FileTypes {
         return FileTypes.WEBP;
       case MimeTypes.IMAGE_BMP:
         return FileTypes.BMP;
+      case MimeTypes.IMAGE_HEIF:
+        return FileTypes.HEIF;
       default:
         return FileTypes.UNKNOWN;
     }
@@ -314,6 +321,8 @@ public final class FileTypes {
       return FileTypes.WEBP;
     } else if (filename.endsWith(EXTENSION_BMP) || filename.endsWith(EXTENSION_DIB)) {
       return FileTypes.BMP;
+    } else if (filename.endsWith(EXTENSION_HEIC)) {
+      return FileTypes.HEIF;
     } else {
       return FileTypes.UNKNOWN;
     }