+++ libraries/extractor/src/main/java/androidx/media3/extractor/DefaultExtractorsFactory.java
@@ -17,6 +17,8 @@ package androidx.media3.extractor;
 
 import static androidx.media3.common.FileTypes.inferFileTypeFromResponseHeaders;
 import static androidx.media3.common.FileTypes.inferFileTypeFromUri;
+import static androidx.media3.extractor.mp4.Mp4Extractor.FLAG_READ_MOTION_PHOTO_METADATA;
+import static androidx.media3.extractor.mp4.Mp4Extractor.FLAG_READ_SEF_DATA;
 
 import android.net.Uri;
 import androidx.annotation.GuardedBy;
@@ -33,6 +35,7 @@ import androidx.media3.extractor.avi.AviExtractor;
 import androidx.media3.extractor.bmp.BmpExtractor;
 import androidx.media3.extractor.flac.FlacExtractor;
 import androidx.media3.extractor.flv.FlvExtractor;
+import androidx.media3.extractor.heif.HeifExtractor;
 import androidx.media3.extractor.jpeg.JpegExtractor;
 import androidx.media3.extractor.mkv.MatroskaExtractor;
 import androidx.media3.extractor.mp3.Mp3Extractor;
@@ -91,6 +94,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
  *   <li>PNG ({@link PngExtractor})
  *   <li>WEBP ({@link WebpExtractor})
  *   <li>BMP ({@link BmpExtractor})
+ *   <li>HEIF ({@link HeifExtractor})
  *   <li>MIDI, if available, the MIDI extension's {@code androidx.media3.decoder.midi.MidiExtractor}
  *       is used.
  * </ul>
@@ -124,7 +128,8 @@ public final class DefaultExtractorsFactory implements ExtractorsFactory {
         FileTypes.JPEG,
         FileTypes.PNG,
         FileTypes.WEBP,
-        FileTypes.BMP
+        FileTypes.BMP,
+        FileTypes.HEIF
       };
 
   private static final ExtensionLoader FLAC_EXTENSION_LOADER =
@@ -523,6 +528,12 @@ public final class DefaultExtractorsFactory implements ExtractorsFactory {
         break;
       case FileTypes.BMP:
         extractors.add(new BmpExtractor());
+        break;
+      case FileTypes.HEIF:
+        if ((mp4Flags & FLAG_READ_MOTION_PHOTO_METADATA) == 0
+            && (mp4Flags & FLAG_READ_SEF_DATA) == 0) {
+          extractors.add(new HeifExtractor());
+        }
         break;
       case FileTypes.WEBVTT:
       case FileTypes.UNKNOWN: